+++ halogen/types.py
@@ -1,6 +1,8 @@
 """Halogen basic types."""
 import datetime
 import decimal
+import enum
+from typing import Union, Optional, Any
 
 import dateutil.parser
 import isodate
@@ -8,6 +10,7 @@ import pytz
 import six
 
 from .exceptions import ValidationError
+from .schema import _Schema
 
 
 class Type(object):
@@ -54,22 +57,27 @@ class List(Type):
         :param item_type: Item type or schema.
         :param allow_scalar: Automatically convert scalar value to the list.
         """
-        super(List, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self.item_type = item_type or Type()
         self.allow_scalar = allow_scalar
 
     def serialize(self, value, **kwargs):
         """Serialize every item of the list."""
-        return [self.item_type.serialize(val, **kwargs) for val in value]
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+        return super().serialize([self.item_type.serialize(val, **kwargs) for val in value], **kwargs)
 
     def deserialize(self, value, **kwargs):
         """Deserialize every item of the list."""
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
         if not isinstance(value, (list, tuple)):
             if self.allow_scalar:
                 value = [value]
             else:
                 raise ValidationError('"{}" is not a list'.format(value))
-        value = super(List, self).deserialize(value)
+        value = super().deserialize(value)
         result = []
         errors = []
 
@@ -81,7 +89,7 @@ class List(Type):
                 errors.append(exc)
         if errors:
             raise ValidationError(errors)
-        return result
+        return super().deserialize(result, **kwargs)
 
 
 class ISODateTime(Type):
@@ -91,7 +99,10 @@ class ISODateTime(Type):
     message = u"'{val}' is not a valid ISO-8601 datetime"
 
     def serialize(self, value, **kwargs):
-        return value.isoformat() if value else None
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        return super().serialize(value.isoformat(), **kwargs)
 
     def deserialize(self, value, **kwargs):
         value = value() if callable(value) else value
@@ -101,7 +112,7 @@ class ISODateTime(Type):
         except (isodate.ISO8601Error, ValueError):
             raise ValueError(self.message.format(val=value))
 
-        return super(ISODateTime, self).deserialize(value)
+        return super().deserialize(value)
 
 
 class ISOUTCDateTime(Type):
@@ -119,9 +130,15 @@ class ISOUTCDateTime(Type):
         return value.isoformat().replace("+00:00", "Z")
 
     def serialize(self, value, **kwargs):
-        return self.format_as_utc(value) if value else None
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        return super().serialize(self.format_as_utc(value), **kwargs)
 
     def deserialize(self, value, **kwargs):
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
         value = value() if callable(value) else value
         try:
             dateutil.parser.parse(value)
@@ -129,7 +146,7 @@ class ISOUTCDateTime(Type):
         except (isodate.ISO8601Error, ValueError):
             raise ValueError(self.message.format(val=value))
 
-        return super(ISOUTCDateTime, self).deserialize(value)
+        return super().deserialize(value)
 
 
 class ISOUTCDate(ISOUTCDateTime):
@@ -143,51 +160,65 @@ class String(Type):
     """String schema type."""
 
     def serialize(self, value, **kwargs):
-        if value:
-            return six.text_type(super(String, self).serialize(value))
-        return ""
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+        return super().serialize(six.text_type(value), **kwargs)
 
     def deserialize(self, value, **kwargs):
-        if value is not None:
-            return super(String, self).deserialize(six.text_type(value))
-        return ""
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+        return super().deserialize(six.text_type(value), **kwargs)
 
 
 class Int(Type):
     """Int schema type."""
 
     def serialize(self, value, **kwargs):
-        return int(value) if value is not None else None
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+        return super().serialize(int(value), **kwargs)
 
     def deserialize(self, value, **kwargs):
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
         try:
-            value = int(value) if value is not None else None
+            value = int(value)
         except ValueError:
             raise ValueError(u"'{val}' is not an integer".format(val=value))
-        return super(Int, self).deserialize(value, **kwargs)
+        return super().deserialize(value, **kwargs)
 
 
 class Boolean(Type):
     """Boolean schema type."""
 
     def serialize(self, value, **kwargs):
-        return bool(value) if value is not None else None
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+        return super().serialize(bool(value), **kwargs)
 
-    def deserialize(self, value, **kwargs):
-        try:
-            value = int(value)
-            value = bool(value) if value == 1 or value == 0 else None
-        except ValueError:
+    def deserialize(self, value: Union[str, int, bool, None], **kwargs):
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        if isinstance(value, int):
+            if value not in (0, 1):
+                raise ValueError("'{val}' is not 1 or 0".format(val=value))
+            value = bool(value)
+
+        if isinstance(value, str):
             str_value = str(value).lower()
+            if str_value not in ("0", "1", "true", "false"):
+                raise ValueError("'{val}' is not 1, 0, true or false".format(val=value))
+
             if str_value == "true":
                 value = True
             elif str_value == "false":
                 value = False
+            else:
+                value = int(str_value)
+                value = bool(value)
 
-        if not isinstance(value, bool):
-            raise ValueError("'{val}' is not an 1 or 0 and true or false".format(val=value))
-
-        return value
+        return super().deserialize(value, **kwargs)
 
 
 class Amount(Type):
@@ -203,9 +234,9 @@ class Amount(Type):
         """
         self.currencies = currencies
         self.amount_class = amount_class
-        super(Amount, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
-    def amount_object_to_dict(self, amount):
+    def amount_object_to_dict(self, amount) -> dict[str, str]:
         """Return the dictionary representation of an Amount object.
 
         Amount object must have amount and currency properties and as_tuple method which will return (currency, amount)
@@ -235,8 +266,9 @@ class Amount(Type):
         :return: Converted amount.
         """
         if value is None:
-            return None
-        return self.amount_object_to_dict(value)
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        return super().serialize(self.amount_object_to_dict(value), **kwargs)
 
     def deserialize(self, value, **kwargs):
         """Deserialize the amount.
@@ -249,7 +281,7 @@ class Amount(Type):
         :raises ValidationError: when amount has more than 2 decimal places
         """
         if value is None:
-            return None
+            raise ValueError("None passed, use Nullable type for nullable values")
 
         if isinstance(value, six.string_types):
             currency = value[:3]
@@ -274,22 +306,61 @@ class Amount(Type):
             raise ValueError(u"'{amount}' has more than 2 decimal places.".format(amount=amount))
 
         value = self.amount_class(currency=currency, amount=amount)
-        return super(Amount, self).deserialize(value)
+        return super().deserialize(value)
 
 
 class Nullable(Type):
     """Nullable type."""
 
-    def __init__(self, nested_type, *args, **kwargs):
+    def __init__(self, nested_type: Union[type[Type], Type, _Schema], *args, **kwargs):
         self.nested_type = nested_type
-        super(Nullable, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
-    def serialize(self, value, **kwargs):
+    def serialize(self, value: Optional[Any], **kwargs):
         if value is None:
             return None
         return self.nested_type.serialize(value, **kwargs)
 
-    def deserialize(self, value, **kwargs):
+    def deserialize(self, value: Optional[Any], **kwargs):
         if value is None:
             return None
         return self.nested_type.deserialize(value, **kwargs)
+
+
+class Enum(Type):
+    """Enum schema type for enum.Enum."""
+
+    def __init__(
+        self,
+        enum_type: type[enum.Enum],
+        use_values: bool = False,
+        *args,
+        **kwargs,
+    ):
+        super().__init__(*args, **kwargs)
+        if not issubclass(enum_type, enum.Enum):
+            raise TypeError("Must be subclass of enum.Enum.")
+        self.enum_type = enum_type
+        self.use_values = use_values
+
+    def serialize(self, value: Optional[enum.Enum], **kwargs):
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        value = value.value if self.use_values else value.name
+
+        return super().serialize(value, **kwargs)
+
+    def deserialize(self, value: Optional[str], **kwargs):
+        if value is None:
+            raise ValueError("None passed, use Nullable type for nullable values")
+
+        if self.use_values:
+            value = self.enum_type(value)
+        else:
+            try:
+                value = self.enum_type[value]
+            except KeyError:
+                raise ValueError(f"Unknown enum key: {value}")
+
+        return super().deserialize(value, **kwargs)