+++ tests/test_types.py
@@ -1,6 +1,8 @@
 """Test halogen types."""
 import decimal
 import datetime
+import enum
+from typing import Union
 
 import pytz
 import six
@@ -9,8 +11,10 @@ import mock
 import pytest
 
 from pytz import timezone
-from halogen import types
+
+from halogen import types, Schema, Attr, exceptions
 from halogen.exceptions import ValidationError
+from halogen.types import Type
 
 
 def test_type():
@@ -29,12 +33,19 @@ def test_list():
     assert value == type_.deserialize(value)
 
 
-@pytest.mark.parametrize("input", [{"foo": "bar"}, 42, 11.5, True, False, None, "", "foo"])
+@pytest.mark.parametrize("input", [{"foo": "bar"}, 42, 11.5, True, False, "", "foo"])
 def test_list_bad_input(input):
     """Test that the deserialization fails correctly when the input is not a list, and scalars are not allowed"""
     type_ = types.List(allow_scalar=False)
     with pytest.raises(ValidationError, match=".*is not a list"):
         type_.deserialize(input)
+
+
+def test_list_none_input():
+    """Test that the deserialization fails correctly when the input is not a list, and scalars are not allowed"""
+    type_ = types.List(allow_scalar=False)
+    with pytest.raises(ValueError, match="None passed, use Nullable type for nullable values"):
+        type_.deserialize(None)
 
 
 @pytest.mark.parametrize(
@@ -64,9 +75,9 @@ def test_isoutcdatetime():
 def test_isoutcdatetime_bc():
     """Test iso datetime with year before 1900."""
     type_ = types.ISOUTCDateTime()
-    value = datetime.datetime(1800, 1, 1, tzinfo=pytz.timezone("CET"))
-    assert type_.serialize(value) == "1799-12-31T23:00:00Z"
-    assert type_.deserialize("1799-12-31T23:00:00Z") == value.replace(microsecond=0)
+    value = datetime.datetime(1800, 1, 1, tzinfo=pytz.timezone("UTC"))
+    assert type_.serialize(value) == value.strftime("%Y-%m-%dT%H:%M:%SZ")
+    assert type_.deserialize("1800-01-01T00:00:00Z") == value.replace(microsecond=0)
 
 
 @pytest.mark.parametrize("value", ["01.01.1981 11:11:11", "123x3"])
@@ -92,7 +103,8 @@ def test_string():
     type_ = types.String()
     value = object()
     assert type_.serialize(value) == six.text_type(value)
-    assert type_.serialize(None) == ""
+    with pytest.raises(ValueError):
+        type_.serialize(None)
     assert type_.deserialize("") == ""
     assert type_.deserialize("Some") == "Some"
     assert type_.deserialize({"key": "value"}) == "{'key': 'value'}"
@@ -115,10 +127,12 @@ def test_int():
         (True, True, True),
         (1, 1, True),
         (0, 0, False),
+        ("1", 1, True),
+        ("0", 0, False),
         ("true", "1", True),
         ("True", "0", True),
         ("false", 0, False),
-        ("False", 0.0, False),
+        ("False", 0, False),
     ],
 )
 def test_boolean(value, clean_value, expected):
@@ -128,11 +142,14 @@ def test_boolean(value, clean_value, expected):
     assert type_.deserialize(value) is expected
 
 
-def test_boolean_invalid():
+@pytest.mark.parametrize(
+    "value", ["not-int", None, "2"],
+)
+def test_boolean_invalid(value):
     """Test boolean type when value provided is not valid."""
     type_ = types.Boolean()
     with pytest.raises(ValueError):
-        type_.deserialize("not-int")
+        type_.deserialize(value)
 
 
 @pytest.mark.parametrize(
@@ -157,7 +174,6 @@ def test_amount_invalid(value, expected):
     ["value", "expected"],
     [
         ("EUR1.22", {"currency": "EUR", "amount": decimal.Decimal("1.22")}),
-        (None, None),
         ({"amount": "12.34", "currency": "EUR"}, {"currency": "EUR", "amount": decimal.Decimal("12.34")}),
         ("EUR12.34", {"currency": "EUR", "amount": decimal.Decimal("12.34")}),
     ],
@@ -166,6 +182,13 @@ def test_amount_valid(value, expected):
     """Test amount type with valid input."""
     type_ = types.Amount(currencies=["EUR"], amount_class=dict)
     assert type_.deserialize(value) == expected
+
+
+@pytest.mark.parametrize("value", [None])
+def test_amount_invalid(value):
+    type_ = types.Amount(currencies=["EUR"], amount_class=dict)
+    with pytest.raises(ValueError):
+        type_.deserialize(value)
 
 
 def test_amount_serialize():
@@ -179,7 +202,8 @@ def test_amount_serialize():
 
     amount.as_quantized().as_tuple.return_value = ("EUR", 11.23)
     assert type_.serialize(amount) == {"amount": "11.23", "currency": "EUR"}
-    assert type_.serialize(None) is None
+    with pytest.raises(ValueError):
+        type_.serialize(None)
 
 
 def test_nullable_type():
@@ -196,3 +220,90 @@ def test_nullable_type():
 
     assert nullable.deserialize("test") == "deserialize"
     assert nullable.deserialize(None) is None
+
+
+@pytest.mark.parametrize(
+    ["type", "value", "result"],
+    [
+        (types.Boolean(), {"required": True, "optional": True, "non_nullable": True}, {"required": True, "optional": True, "non_nullable": True}),
+        (types.Boolean(), {"required": None, "optional": None, "non_nullable": True}, {"required": None, "optional": None, "non_nullable": True}),
+        (types.Boolean(), {"required": None, "non_nullable": True}, {"required": None, "non_nullable": True}),
+    ],
+)
+def test_nullable_attribute(type: Type, value: dict, result: dict):
+
+    class FooBar(Schema):
+        required = Attr(types.Nullable(type), required=True)
+        optional = Attr(types.Nullable(type), required=False)
+        non_nullable = Attr(type, required=False)
+
+    assert FooBar.deserialize(value) == result
+
+
+@pytest.mark.parametrize(
+    ["type", "value"],
+    [
+        (types.Boolean(), {"non_nullable": True}),
+        (types.Boolean(), {"required": True, "non_nullable": None}),
+    ],
+)
+def test_failing_nullable_attribute(type: Type, value: dict):
+
+    class FooBar(Schema):
+        required = Attr(types.Nullable(type), required=True)
+        optional = Attr(types.Nullable(type), required=False)
+        non_nullable = Attr(type, required=False)
+
+    with pytest.raises(ValidationError):
+        FooBar.deserialize(value)
+
+
+@pytest.mark.parametrize(
+    ["value", "use_values"],
+    [
+        (1, True),
+        ("FOO", False),
+        ("BAR", False),
+    ],
+)
+def test_enum(value: Union[int, str, None], use_values: bool):
+    class TestEnum(enum.Enum):
+        FOO = 1
+        BAR = 2
+
+    type_ = types.Enum(TestEnum, use_values=use_values)
+
+    assert isinstance(type_.deserialize(value), TestEnum)
+
+
+@pytest.mark.parametrize(
+    ["value", "use_values", "expected_error"],
+    [
+        (1, False, "Unknown enum key"),
+        (3, True, "is not a valid"),
+        ("FOO", True, "is not a valid"),
+        ("BAR", True, "is not a valid"),
+        ("NONEXISTENT", False, "Unknown enum key"),
+        (1.5, True, "is not a valid")
+    ],
+)
+def test_invalid_enum(value: Union[int, str, float], use_values: bool, expected_error: str):
+    class TestEnum(enum.Enum):
+        FOO = 1
+        BAR = 2
+
+    type_ = types.Enum(TestEnum, use_values=use_values)
+
+    with pytest.raises(ValueError, match=expected_error):
+        type_.deserialize(value)
+
+
+def test_nullable_enum():
+    class TestEnum(enum.Enum):
+        FOO = 1
+        BAR = 2
+
+    type_ = types.Enum(TestEnum, use_values=False)
+
+    with pytest.raises(ValueError):
+        type_.serialize(None)